Update wheelchair and toilets:wheelchair tag in any case
The previous solution did not take into account that it could happen
that a user possibly updates both the wheelchair *and* the
toilets:wheelchair tag at the same time. So it only saved one of them.
Now it just checks if the wheelchair/toilets:wheelchair got a new value
and updates it in the wheelmap database accordingly