Read map data from cookies and paste it into support mail
Getting our hands on latitude, longitude and zoom level is not as easy
as with the other data. One approach would be to these from the URL as
they're passed in as query params however query params are not part of
the referer.
We now grab these from the cookie since the Ember frontend persists
these settings there.