fix misplaced map marker
It can happen that we receive multiple results for a certain address.
For example `Prenzlauer Allee 23` returns more than one match. Before we
took just the first match and assumed it is the correct one. Now we get
back more results from Photon and start searching for the correct one
manually