Add node_type to short node GeoJSON
This allows the new frontend to use the node_type attribute for map markers.

Background: The new frontend includes a mechanism to categorize PoIs using synonymes. When developing accessibility.cloud, Holger, Thomas, Katharina and I unified the various systems we found. The new frontend displays subtly different icons for that reason, and to do this correctly, it needs the exact node_type info for each PoI.

Tested locally. Could somebody merge / deploy this? :)